requirements.txt
langchain
faiss-cpu
sentence-transformers
pypdf
beautifulsoup4
requests
streamlit
fastapi
uvicorn
ollama
tqdm
numpy